Large changes, see description.
Added a router path for checkoutPage which is a child of the landing page but can only be accessed via the checkout button in the shopping cart.

Created a function for creating  Google Forms, however this might be useless and will potentially be deleted later.

Will use local form instead to push data into a DB.

Orderform jsx and css added and creatred.