Large update (see description)
added .env.local to .gitignore

fixed import path for orderprovider.

Ensured cartItem components for the shopping cart page were mapped to have a key = index they are generated in.

Changed OrderFormjsx to require setUserDetails state and pushOrderRequest function

Subtotalchecker refactored to not take handleCheckout as a paramater.

Instead NTCP and isSCP credirect to checkout page and check current page user is on.

Deleted obsolete testData.

Refactored checkoutTotal.js helper function. Removed functions from subtotalChecker.jsx and moved fn into checkoutTotal.

updated the useFireBaseOrders and useShoppingCart hook and path.

CheckoutPage now supports fn "pushOrderRequest".

This fn checks if there are any items in the cart and generates an object called formatted items, which is pushed as items to the db, subtotal, which subtotals the value of the cart, and userDetails, which is a state that is pulled from an order form.

Handlecheckout fn renamed to navigateToCheckoutPage